Change mana value to "real" and "effective"
"Real" mana value is the MV as defined by the rules of the card, which
is relevant for effects that look in, e.g., a player's hand or library.
"Effective" mana value is the per-face MV while on the stack or
battlefield, which is relevant for, e.g., effects that care about spells
you cast or permanents you control.